# openapi.yaml
openapi: 3.0.3
info:
  title: JF DB API
  version: 1.0.0
  description: API for All Out events, participants, maps, leaderboards, and Steam authentication.
servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: "3000"
tags:
  - name: All Out events
  - name: All Out leaderboards
  - name: Steam authentication
paths:
  /all-out/events:
    get:
      tags: [All Out events]
      summary: List All Out events
      operationId: listAllOutEvents
      responses:
        "200":
          description: Event previews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AllOutEventPreview'
    post:
      tags: [All Out events]
      summary: Create an All Out event
      operationId: createAllOutEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAllOutEvent'
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllOutEventDetails'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [All Out events]
      summary: Update an All Out event
      operationId: updateAllOutEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAllOutEvent'
      responses:
        "204":
          description: Event updated
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [All Out events]
      summary: Get All Out event details
      operationId: getAllOutEvent
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllOutEventDetails'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}/participants:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [All Out events]
      summary: List event participants
      operationId: listAllOutParticipants
      responses:
        "200":
          description: Steam users registered for the event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SteamUser'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}/is-registered:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [All Out events]
      summary: Check whether the authenticated user is registered
      operationId: checkAllOutRegistration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Registration status
          content:
            application/json:
              schema:
                type: object
                required: [isRegistered]
                properties:
                  isRegistered:
                    type: boolean
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}/maps/stage-1:
    parameters:
      - $ref: '#/components/parameters/EventId'
    post:
      tags: [All Out events]
      summary: Replace stage 1 maps
      operationId: replaceStage1Maps
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateTimeLimitedEventMap'
      responses:
        "201":
          description: Updated stage 1 maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeLimitedEventMap'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}/maps/stage-2:
    parameters:
      - $ref: '#/components/parameters/EventId'
    post:
      tags: [All Out events]
      summary: Replace stage 2 maps
      operationId: replaceStage2Maps
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventMaps'
      responses:
        "201":
          description: Updated stage 2 maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventMap'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}/maps/stage-3:
    parameters:
      - $ref: '#/components/parameters/EventId'
    post:
      tags: [All Out events]
      summary: Replace stage 3 maps
      operationId: replaceStage3Maps
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventMaps'
      responses:
        "201":
          description: Updated stage 3 maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventMap'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/events/{eventId}/register:
    parameters:
      - $ref: '#/components/parameters/EventId'
    post:
      tags: [All Out events]
      summary: Register the authenticated user for an event
      operationId: registerForAllOutEvent
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User registered
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/leaderboard/stage-1:
    get:
      tags: [All Out leaderboards]
      summary: Get the stage 1 leaderboard
      operationId: getStage1Leaderboard
      parameters:
        - $ref: '#/components/parameters/MapId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Order'
      responses:
        "200":
          description: Stage 1 leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardItem'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/leaderboard/stage-2:
    get:
      tags: [All Out leaderboards]
      summary: Get the stage 2 leaderboard
      operationId: getStage2Leaderboard
      parameters:
        - $ref: '#/components/parameters/MapId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Order'
      responses:
        "200":
          description: Stage 2 leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LapLeaderboardItem'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
  /all-out/leaderboard/stage-3:
    get:
      tags: [All Out leaderboards]
      summary: Get the stage 3 leaderboard
      operationId: getStage3Leaderboard
      parameters:
        - $ref: '#/components/parameters/MapId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Order'
      responses:
        "200":
          description: Stage 3 leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardItem'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
  /steam-auth/init:
    get:
      tags: [Steam authentication]
      summary: Start Steam authentication
      operationId: initSteamAuthentication
      responses:
        "302":
          description: Redirect to Steam OpenID login
          headers:
            Location:
              schema:
                type: string
                format: uri
  /steam-auth/callback:
    get:
      tags: [Steam authentication]
      summary: Complete Steam authentication
      operationId: completeSteamAuthentication
      parameters:
        - name: openid.ns
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: openid.mode
          in: query
          required: true
          schema:
            type: string
        - name: openid.op_endpoint
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: openid.claimed_id
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: openid.identity
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: openid.return_to
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: openid.response_nonce
          in: query
          required: true
          schema:
            type: string
        - name: openid.assoc_handle
          in: query
          required: true
          schema:
            type: string
        - name: openid.signed
          in: query
          required: true
          schema:
            type: string
        - name: openid.sig
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: JWT and authenticated Steam user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SteamAuthResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      description: Positive event identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
    MapId:
      name: mapId
      in: query
      required: true
      description: Positive map identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
    Page:
      name: page
      in: query
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
    PageSize:
      name: pageSize
      in: query
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
    Order:
      name: order
      in: query
      required: true
      schema:
        type: string
        enum: [ASC, DESC]
  responses:
    BadRequest:
      description: Invalid path, query, or request body data.
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        text/plain:
          schema:
            type: string
            example: Unauthorized
    Forbidden:
      description: Authenticated user is not an administrator.
      content:
        text/plain:
          schema:
            type: string
            example: Forbidden
    NotFound:
      description: Requested resource was not found.
      content:
        text/plain:
          schema:
            type: string
    InternalServerError:
      description: Unexpected server or Steam provider error.
      content:
        text/plain:
          schema:
            type: string
  schemas:
    DateTime:
      type: string
      format: date-time
    AllOutEventPreview:
      type: object
      required: [id, start, end]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        start:
          $ref: '#/components/schemas/DateTime'
        end:
          $ref: '#/components/schemas/DateTime'
    EventMap:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
    TimeLimitedEventMap:
      allOf:
        - $ref: '#/components/schemas/EventMap'
        - type: object
          required: [timeLimit]
          properties:
            timeLimit:
              type: number
              minimum: 0
              exclusiveMinimum: true
    AllOutStage:
      type: object
      required: [description, start, end, maps]
      properties:
        description:
          type: string
        start:
          $ref: '#/components/schemas/DateTime'
        end:
          $ref: '#/components/schemas/DateTime'
        maps:
          type: array
          items:
            $ref: '#/components/schemas/EventMap'
    AllOutStage1:
      allOf:
        - $ref: '#/components/schemas/AllOutStage'
        - type: object
          properties:
            maps:
              type: array
              items:
                $ref: '#/components/schemas/TimeLimitedEventMap'
    AllOutEventDetails:
      type: object
      required: [id, description, stage1, stage2, stage3]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string
        stage1:
          $ref: '#/components/schemas/AllOutStage1'
        stage2:
          $ref: '#/components/schemas/AllOutStage'
        stage3:
          $ref: '#/components/schemas/AllOutStage'
    AllOutEventStageInput:
      type: object
      required: [start, end]
      properties:
        start:
          $ref: '#/components/schemas/DateTime'
        end:
          $ref: '#/components/schemas/DateTime'
        description:
          type: string
    CreateAllOutEvent:
      type: object
      required: [description, stage1, stage2, stage3]
      properties:
        description:
          type: string
          minLength: 1
        stage1:
          $ref: '#/components/schemas/AllOutEventStageInput'
        stage2:
          $ref: '#/components/schemas/AllOutEventStageInput'
        stage3:
          $ref: '#/components/schemas/AllOutEventStageInput'
    UpdateAllOutEvent:
      allOf:
        - $ref: '#/components/schemas/CreateAllOutEvent'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              minimum: 1
    CreateEventMap:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
    CreateEventMaps:
      type: array
      items:
        $ref: '#/components/schemas/CreateEventMap'
    CreateTimeLimitedEventMap:
      allOf:
        - $ref: '#/components/schemas/CreateEventMap'
        - type: object
          required: [timeLimit]
          properties:
            timeLimit:
              type: number
              minimum: 0
              exclusiveMinimum: true
    SteamUser:
      type: object
      required: [steamId64, name, avatar]
      properties:
        steamId64:
          type: string
        name:
          type: string
        avatar:
          type: string
          format: uri
    LeaderboardItem:
      type: object
      required: [id, user, prSeconds, prTimestamp]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        user:
          $ref: '#/components/schemas/SteamUser'
        prSeconds:
          type: number
        prTimestamp:
          $ref: '#/components/schemas/DateTime'
    LapLeaderboardItem:
      allOf:
        - $ref: '#/components/schemas/LeaderboardItem'
        - type: object
          required: [lapCount, lastLapTimestamp]
          properties:
            lapCount:
              type: integer
            lastLapTimestamp:
              $ref: '#/components/schemas/DateTime'
    SteamAuthResponse:
      type: object
      required: [token, user, role]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/SteamUser'
        role:
          type: string
          enum: [admin, user]